+++ src/funcs/firewallsGetAllFirewallAssignments.ts
@@ -29,7 +29,7 @@ import { Result } from "../types/fp.js";
  * List firewall assignments
  *
  * @remarks
- * Returns a list of all servers and virtual machines assigned to one or more firewalls.
+ * Returns a list of all servers and virtual machines assigned to a firewall.
  */
 export function firewallsGetAllFirewallAssignments(
   client$: LatitudeshCore,